requirements.txt
attr==0.3.2
ConfigParser==6.0.0
contextlib2==21.6.0
cryptography==42.0.3
docutils==0.20.1
filelock==3.13.1
gensim==4.3.2
HTMLParser==0.0.2
importlib_metadata==7.0.1
ipython==8.21.0
ipywidgets==8.1.2
Jinja2==3.1.3
jnius==1.1.0
keyring==24.3.0
langdetect==1.0.9
matplotlib==3.8.3
mock==5.1.0
nltk==3.8.1
numpy==1.26.4
pandas==2.2.0
Pillow==10.2.0
plotly==5.19.0
protobuf==4.25.3
pyLDAvis==3.4.1
pyOpenSSL==24.0.0
railroad==0.5.0
redis==5.0.1
scikit_learn==1.4.1.post1
seaborn==0.13.2
snscrape==0.7.0.20230622
Sphinx==7.2.6
streamlit==1.31.1
thread==0.1.3
tornado==6.4
Unidecode==1.3.8
urllib3_secure_extra==0.1.0
wordcloud==1.9.3
xmlrpclib==1.0.1
